Aggiunto controllo che permette l'eliminazione di una certa riga tra i contract month recap, solo se l'entità è persistent

Closes #732

See merge request epas/epas!643

+++ app/manager/ContractManager.java
@@ -221,7 +221,9 @@ public class ContractManager {
       // TODO: anche quelli sulle ferie quando ci saranno
       for (ContractMonthRecap cmr : contract.contractMonthRecaps) {
         if (!yearMonthFrom.isAfter(new YearMonth(cmr.year, cmr.month))) {
-          cmr.delete();
+          if (cmr.isPersistent()) {
+            cmr.delete();
+          }          
         }
       }
